test: align external review gate api boundaries
