+++ backend/tests/integration/test_external_review_gate_http_api/__init__.py
@@ -1,0 +1,1 @@
+"""ExternalReviewGate HTTP API integration test package."""
